requirements.txt
fastapi 
uvicorn 
httpx 
pyjwt 
python-jose[cryptography] 
python-dotenv